fix: Remove state_key: null from search results
Seshat stores events as-is, but some clients may include "state_key": null
for non-state events. The matrix-js-sdk checks for the presence of state_key
to determine if an event is a state event, so having state_key: null causes
non-state events to be misidentified as state events, breaking search result
rendering in Element.

This fix removes state_key: null from:
- Main search result events
- Context events (events_before/events_after)